syscalls/inotify04: Fix cleanup
The cleanup attempted to close() the inotify watch descriptors however
these are _NOT_ file descriptors and are allocated from completely
distinct space.

So in this case the inotify watch descriptors we got were 1 and 2 so we
closed the stdout and stderr at the end of the test which still made the
test pass without any error just by a chance.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>