commands/which: Fix wrong word splitting
Space in "pname='pname -i'" causes wrong splitting in test 3 and 4
as the default IFS (variable used when doing word splitting) is space.

Input
"pname='pname -i'" "$PWD/pname"

is wrongly splitted as:
0: pname='pname
1: -i'
2: $PWD/pname

That produced error:
    grep: invalid option -- '''
    Usage: grep [OPTION]... PATTERN [FILE]...
    Try 'grep --help' for more information.
     4 TPASS: 'alias | which --read-alias pname' passed.
    grep: invalid option -- '''
    Usage: grep [OPTION]... PATTERN [FILE]...
    Try 'grep --help' for more information.
     5 TPASS: 'alias | which -i pname' passed.

There is emulation of two dimensional array and both would be broken by
this space, let's fix that by using two IFS: '^' for top level (all of
these must be passed) and second level '|' (any of these must be find).

This affected only 'which' command which supports --read-alias options
(IMHO only 'GNU which') and the result was the same as after fix, which
hided this error.

+ Add -F for grep to have '.' as real dot and not a wildcard.

Signed-off-by: Petr Vorel <pvorel@suse.cz>