netstress: handle pmtu update on EMSGSIZE error
After the error we should get a new route cache with updated
pmtu, resend datagram again using the new one.

The exact number of errors during the test may depend on the
route cache expire time, by default it is 600 sec:
  net.ipv4.route.mtu_expires
  net.ipv6.route.mtu_expires

and the total test time. Therefore it is difficult to anticipate
what value we would need, especially for the server that don't
have the requests number parameter. Set it to 10, it should be
enough for most of the test scenarios.

Signed-off-by: Alexey Kodanev <alexey.kodanev@oracle.com>