syscalls/perf_event_open02: Fix failures
The testcase was failing randomly, it could be reproduced easily when
the machine is under load. To reproduce the issue just run a few
dd if=/dev/zero of=/dev/null to saturate your CPUs.

It has been sugessted that the reason for the failure are rounding
errors caused by a frequent preemption. I haven't got a definitive
answer from kernel devs for this but it's true that changing the process
pritority to realtime SCHED_FIFO for the measurement does fix the issue.

So we either delete the test or apply this patch.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>