lib/tst_device: Fill a file in megabyte sized chunks
A zero-filled file is used as backing store when creating a loopback
device. The zero-fill operations is done in 1KB sized chunks for
file-size specified in megabytes. This leads to a large number of
write calls which take long time during test setup phase in certain
environments.

Move to using 1MB sized chunks to reduce the number of write syscalls
by a few order of magnitude.

Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>