testcase: gethostid: Check if hostid is available
On some embedded systems, hostid may not be installed.
Because of that, the test will fail with:

gethostid01    1  TFAIL  :  gethostid01.c:171: fgets failed

The test should return a TCONF instead.
This commit uses tst_get_path function to check if "hostid"
is available in $PATH. Otherwise, it will return a TCONF.

Signed-off-by: Mylène Josserand <mylene.josserand@bootlin.com>
Acked-by: Jan Stancek <jstancek@redhat.com>