device-drivers/acpi: skip the test if ACPI is disabled
Calling functions such as acpi_get_devices() on a system with ACPI
disabled (for example, a virtual machine) leads to a kernel panic.

Earlier there was a related discussion on linux-acpi:

https://www.spinics.net/lists/linux-acpi/msg79419.html

and they came to the conclusion that the caller of the API should
check if ACPI is enabled, i.e. the test needs an update.

Suggested-by: Rob Gardner <rob.gardner@oracle.com>
Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>