realtime/run.sh: Append default number of iteration
Append default number of iterations (1) at the end of the array
otherwise the when executed as:

$ ./testcases/realtime/run.sh -t all

we get

./run.sh: line 87: 0 + : syntax error: operand expected (error token is "+ ")

In the run.sh we do create an array that is supposedly filled with
testname, number_of_iterations, testname, number_of_iterations, ...

And the problem is that the loop fails to add default number of
iterations (1) for the last testname, unless specified explicitly with
-l option.

When we pass -t foo -t bar, the array contains {foo, 1, bar} instead of
{foo, 1, bar, 1} hence the error.

It works fine if we specify the number of iterations explicitly with -l
though as -t foo -t bar -l 1.

So the best fix seems to be appending 1 to the array at the end of the
loop, it will add one more useless number to the array in a case that
number of iterations is specified explicitly but other than that will
not do any harm.

Fixes #132

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>