lib: Fix result propagation after exec() + tst_reinit()
The result pointer wasn't initialized in tst_reinit(), which haven't
allowed for test result propagation to the main test process and there
is actually no reason not to do that. So this commit initializes the
results pointer in tst_reinit() and also adds a child_needs_reinit flag
to tst_test structure, which causes the library not to unlink the IPC
file after it has been mapped and also inserts LTP_IPC_PATH to the
environment.

This commit also adds a test that also shows the usage of the API and
documentation.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
CC: Eddie Horng <eddiehorng.tw@gmail.com>
Acked-by: Jan Stancek <jstancek@redhat.com>