open_posix_testsuite/mmap_11-4: Skip this test on tmpfs
According to mmap(2) manpage, when you write data to such partial
page after the end of the object, the data stays in the page cache
even after the file is closed and unmapped, subsequent mappings
always see the modified content on tmpfs.  We just skip this test
when running it on tmpfs.

Signed-off-by: Xiao Yang <yangx.jy@cn.fujitsu.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>