syscalls/setxattr: fix iterations and missing coverage
During issue #274 resolution, the following was observed:

setxattr(2) tests are not considering iterations on files that are not
re-created in between subsequent iterations. This leads to false results
due to existing attributes not cleaned in between the calls.

setxattr(2) tests were not considering 2 cases, now added:
  - new attr with a key with length 0
  - new attr with a key that is NULL

This commit makes setxattr(2) tests to behave like newly created
fsetxattr(2) tests.

Signed-off-by: Rafael David Tinoco <rafael.tinoco@linaro.org>
Acked-by: Cyril Hrubis <chrubis@suse.cz>