syscalls/statx04: Add new test.
This test will check if statx() reports attribute flags by adding
FS_COMPR_FL, FS_APPEND_FL, FS_IMMUTABLE_FL and FS_NODUMP_FL to a
direcotry and checks that the flags are reported correctly and also
tests a directory with no flags.

The test is executed against all filesystems, since most of the kernel
code that implements this functionality is filesystem specific.

* The ntfs3g fuse driver reports wrong errno for unimplemented ioctls()
  EINVAL while every other filesystem reports ENOTTY correctly

* The test fails for ext2 filesystem mounted by kernel ext2 driver,
  which is confirmed to be kernel bug that will be fixed later on

Signed-off-by: Tarun.T.U <tarun@zilogic.com>
Signed-off-by: Vaishnavi.D <vaishnavi.d@zilogic.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>