syscalls/membarrier: Add membarrier() initial test
Fixes: #265

Initial test for membarrier() syscall. It tests all existing membarrier
"commands" (or features), including the need (or not) for previous
registration for the call to work.

Some features did not exist in older kernels and that is covered by
skipping some calls, flagging test as skipped & okay, and forcing
others, making sure that return codes and errno are set right in those
cases.

Tests are done in a child process due to inexistent kernel interface to
"unregister" the process from being affected by membarrier() call.

Signed-off-by: Rafael David Tinoco <rafael.tinoco@linaro.org>
Signed-off-by: Jan Stancek <jstancek@redhat.com>