fzsync: Simplify API with start/end race calls and limit exec time
Make it so the test writer simply has to mark the beginning and end of a
race in each thread. Previously they must choose where to call a delay
and where to call a timestamp function which is used to calculate the
delay as well as the update function.

Along with the new API a randomised delay has been introduced which
helps hit some race conditions (see doc comments).

Also use the tst_timer library to check how long the main test loop has
been running for and break the loop if it exceeds (60 * LTP_TIMEOUT_MUL)
seconds. This prevents an overall test timeout which is reported as a
failure.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.com>
Reviewed-by: Li Wang <liwang@redhat.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>