kernel/syscall/setrlimit05.c: use a legal bad address for mips
This testcase fails on mips32. The process is killed by SIGBUS which
is not as expect.

This is because:
((void *)-1) is not a legal bad address which causes the process
killed by SIGBUG on mips.

'tst_get_bad_addr()' returns an address that should works on mips
and other arches.

Signed-off-by: Hongzhi.Song <hongzhi.song@windriver.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>