Testing statx syscall Timestamp fields
The time before and after the execution of the corresponding
system call is noted.

It is checked whether the time returned by statx lies in this range.

open syscall with O_CREAT flag for birth time.
write syscall for modified time.
read syscall for access time.
chmod syscall for status change time.

Signed-off-by: Subash Ganesan <subash@zilogic.com>
Signed-off-by: Vaishnavi.D <vaishnavi.d@zilogic.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>