msgstress04: correct the logic to use all values in keyarray
There are some problems of the logic to use all values in keyarray. It
fails to run:

$ ./msgstress04 -n 1 -c 1 -l 1

Only the first MSGMNI members of keyarray are initialized. But it will
use the member whose index exceeds MSGMNI in keyarray.

Correct the logic to make sure to use the initialized members of
keyarray.

Signed-off-by: Kai Kang <kai.kang@windriver.com>
Acked-by: Jan Stancek <jstancek@redhat.com>