containers/pidns13: Fix race between threads
The original code was not thread safe and relied on thread ordering to
run correctly. The problem is that the thread id was passed as a pointer
to a variable that has been changed afterwards, which is obviously not
safe.

So now the code simply casts integer to pointer and back so we pass the
actual value rather than pointer to it, which makes the code simpler and
is race free.

This has been originaly reported in:

https://github.com/linux-test-project/ltp/pull/167

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>