unshare01.sh: protect parameters expansion
The script passes around as args some strings with spaces using
positional parameters: these are then in turn used unquoted to declare
local function vars; this usage of 'local' together with the unquoted
expansion is broken and leads to:

unshare_test()
{
        local unshare_opts=$1
        local verify_cmd=$2        <<<----124:
...
case $1 in
 unshare_test "--user" "id -u" "65534";;
....

unshare01.sh: 124: local: -u: bad variable name

This works fine on Bash but not on Dash (despite not being detected
by checkbashism) and it is closely related to the use of 'local';
i.e. removing instead 'local' while leaving the unquoted expansion
solves too.

Protect the positional parameter expansion with double quotes to
prevent globbing and word splitting.

Signed-off-by: Cristian Marussi <cristian.marussi@arm.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>