lib: Add library function for parsing kernel config
This is meant as last resort action for disabling tests if certain
kernel funcitonality was not present, in general case runtime checks are
prefered.

For functionality that can be build as a module tst_check_driver() is
most likely better fit since it will also insert requested kernel module
into kernel if needed.

For newly added syscalls kernel version comparsion and/or checking errno
is prefered.

However in rare cases certain core kernel functionality cannot be
detected in any other way than checking the kernel config, which is
where this API gets into the play.

The path to the kernel config could be specified by KCONFIG_PATH
environment variable, which also takes precedence before the
autodetection that attempts to read the config from known locations.

The required kernel options are passed as an array of strings via the
.needs_kconfigs pointer in the tst_test structure. The purpose of this
is twofold, one is that the test can disable itself at runtime if given
functionality is missing from kernel .config and second is about being
able to propagate this information to the testrunner (this could be done
once we figure out how export the information from the structure to the
test runner) then we can avoid running tests on unsuitable
configurations from the start.

There are two different ways how to specify the required kernel options
either by stating the config name i.e. "CONFIG_FOO" which is considered
present in kernel if set to a certain value and "CONFIG_FOO=bar" which
requires the config option to be strictly set to a desired value.

The code has been also split into parsing phase that simply picks up the
config option values from the config and to evaluation which does the
actual checks and aborts the test if needed. This makes it easier to
expand the domain language for specifying the required options, if we
ever need to match against subset of values this should be reasonably
easy to implement.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Acked-by: Sandeep Patil <sspatil@google.com>
CC: Pengfei Xu <pengfei.xu@intel.com>
CC: automated-testing@yoctoproject.org