commands/logrotate_tests.sh: Fix missing syslog group on some systems
If syslog group doesn't exist on system, su option with it triggers an
'unknow group' error in config file and that caused logrocate command to
skip the config file. We just use syslog group if it actually exists.

This fixes error since commit e547b94 ("return non-zero exit status if a
config file contains an error") in logrotate.

Signed-off-by: Xiao Yang <yangx.jy@cn.fujitsu.com>
[pvorel: split commits and adjusted commit message]
Signed-off-by: Petr Vorel <pvorel@suse.cz>