lib/tst_test.sh: don't call _tst_do_exit() recursively with tst_brk()
With this patch, "safe" functions such as "ROD" and "tst_rhost_run -s"
can be used in a test cleanup function. In case of an error, "safe"
command won't recursively call cleanup function but will only print
the warning message about the failure.

This behavior is consistent with the LTP C library.

Signed-off-by: Alexey Kodanev <alexey.kodanev@oracle.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>