openposix: Remove _XOPEN_SOURCE in strchr/strcpy/strncpy
We have defined _XOPEN_SOURCE to 600 by gcc -D in open_posix_testsuite/CFLAGS，
So defining _XOPEN_SOURCE to different value in test results in a "_XOPEN_SOURCE
redefined" warning.

It seems that all of tests get pass even if _XOPEN_SOURCE is defined to 600
by default, so remove the duplicate _XOPEN_SOURCE in test.

Signed-off-by: Xiao Yang <yangx.jy@cn.fujitsu.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>