Fix `mounted_noatime` on *BSD
I accidentally omitted one of the arguments to a diagnostic printf I
added when I originally added the code. This addresses the issue by
adding the argument for the `path` parameter to the printf call to add
the diagnostics output I originally intended to add.

This does not affect Linux.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>