Fix the dummy declaration for `lio_listio`
The first parameter should be of `struct aiocb *restrict const[restrict]` type,
not `struct aiocb *const[]` type, per the POSIX spec.

Unfortunately this still doesn't cure compilation with gcc 4.8.5 (it claims
there's an issue with `-Wincompatible-pointer-types`), so compiling this with
-Werror seems like a non-starter until the underlying problem is
resolved.

While here, fix whitespace between types and `*` for pointers.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>