tst_test: Add test variants
The test variants are intended for running the test more than once for a
different settings. The indended purpose is to run tests for both libc
wrapper and raw syscall as well as for different syscall variants.

The commit itself adds a test_variant integer that, if set, denotes
number of test variants. The test is then forked and executed
test_variant times each time with different value in global tst_variant
variable.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Steve Muckle <smuckle@google.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>